Remove unused epollet, fix bugs

+++ src/str/str.h
@@ -5,6 +5,6 @@
 #ifndef NFT_DNS_HASH_H
 #define NFT_DNS_HASH_H
 
-unsigned long strhash(const char *str);
+size_t strhash(const char *str);
 
 #endif //NFT_DNS_HASH_H
